feat: refactor schemas by moving orderSchema to a new file and removing unused schemas; update imports accordingly
