+++ src/app/(frontend)/(home)/(auth)/auth/login/form.tsx
@@ -9,7 +9,6 @@ import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useForm } from 'react-hook-form';
 import { Form, FormControl, FormField, FormItem, FormLabel, FormMessage } from '@/components/ui/form';
-import { submitLoginForm } from '@/actions/login';
 import { useEffect, useState, useRef, useTransition } from 'react';
 import { Loader2 } from 'lucide-react';
 import {
