+++ src/app/(frontend)/dashboard/home/@admin/page.tsx
@@ -7,6 +7,7 @@ import { getPayload } from 'payload';
 import { loadSearchParams } from './searchParams';
 import config from '@/payload.config';
 import { SearchParams } from 'nuqs/server';
+import { ServerStatus } from '@/components/ServerStatus';
 
 //* example data:
 // const data: Order[] = [
@@ -405,69 +406,11 @@ export default async function AdminPage({ searchParams }: { searchParams: Promis
 		depth: 2,
 	});
 
-	// const orderData: Order[] = orders.docs.map(order => {
-	// 	const prints = order.prints.map(print => {
-	// 		const baseProps = {
-	// 			id: print.id || '',
-	// 			blockType: print.blockType,
-	// 			quantity: print.quantity,
-	// 			price: print.price,
-	// 			completed: print.completed || false,
-	// 		};
-
-	// 		if (print.blockType === 'shopProduct') {
-	// 			return {
-	// 				...baseProps,
-	// 				blockType: 'shopProduct' as const,
-	// 				product: typeof print.product === 'string' ? print.product : print.product.id,
-	// 			};
-	// 		}
-
-	// 		return {
-	// 			...baseProps,
-	// 			blockType: 'customPrint' as const,
-	// 			model: print.model,
-	// 			printingOptions: {
-	// 				preset:
-	// 					print.printingOptions.preset === null
-	// 						? undefined
-	// 						: typeof print.printingOptions.preset === 'string'
-	// 							? print.printingOptions.preset
-	// 							: print.printingOptions.preset?.id,
-	// 				layerHeight: print.printingOptions.layerHeight || undefined,
-	// 				infill: print.printingOptions.infill || undefined,
-	// 				plastic: print.printingOptions.plastic,
-	// 				colour: print.printingOptions.colour,
-	// 			},
-	// 			time: print.time || undefined,
-	// 			filament: print.filament || undefined,
-	// 		};
-	// 	});
-
-	// 	return {
-	// 		id: order.id,
-	// 		name: order.name,
-	// 		customer: order.customer as { id: string; name: string },
-	// 		shopProducts: prints.filter(p => p.blockType === 'shopProduct').length,
-	// 		customPrints: prints.filter(p => p.blockType === 'customPrint').length,
-	// 		total: order.pricing.total || 0, // technically will never be 0 since a hook controls the value
-	// 		queue: order.queue || 0,
-	// 		currentStatus: order.status.currentStatus,
-	// 		statuses: order.status.statuses || [],
-	// 		comments: order.comments || undefined,
-	// 		createdAt: order.createdAt,
-	// 		payment: order.payment,
-	// 		shipping: order.shipping,
-	// 		shippingAddress: order.shippingAddress,
-	// 		pricing: order.pricing,
-	// 		prints,
-	// 	};
-	// });
-
 	const orderData: ZodOrder[] = orders.docs;
 
 	return (
 		<div className='@container/main flex flex-1 flex-col gap-2'>
+			<ServerStatus></ServerStatus>
 			<div className='flex flex-col gap-4 py-4 md:gap-6 md:py-6'>
 				<SectionCards />
 				<div className='px-4 lg:px-6'>
