--- src/schemas.ts
@@ -1,231 +1,0 @@
-// Generated by ts-to-zod
-import { z } from "zod";
-import { type User, type Order, type Quote } from "./types";
-
-export const presetSchema = z.object({
-    id: z.string(),
-    name: z.string(),
-    description: z.string().optional().nullable(),
-    bambulabName: z.string().optional().nullable(),
-    updatedAt: z.string(),
-    createdAt: z.string()
-});
-
-export const mediaSchema = z.object({
-    id: z.string(),
-    alt: z.string(),
-    updatedAt: z.string(),
-    createdAt: z.string(),
-    url: z.string().optional().nullable(),
-    thumbnailURL: z.string().optional().nullable(),
-    filename: z.string().optional().nullable(),
-    mimeType: z.string().optional().nullable(),
-    filesize: z.number().optional().nullable(),
-    width: z.number().optional().nullable(),
-    height: z.number().optional().nullable(),
-    focalX: z.number().optional().nullable(),
-    focalY: z.number().optional().nullable()
-});
-
-export const filamentSchema = z.object({
-    id: z.string(),
-    name: z.string(),
-    data: z.union([z.record(z.string(), z.unknown()), z.array(z.unknown()), z.string(), z.number(), z.boolean()]).nullable(),
-    updatedAt: z.string(),
-    createdAt: z.string()
-});
-
-export const productSchema = z.object({
-    id: z.string(),
-    name: z.string(),
-    description: z.string(),
-    pictures: z.array(z.union([z.string(), mediaSchema])),
-    price: z.number(),
-    time: z.string(),
-    orders: z.number().optional().nullable(),
-    printingOptions: z.object({
-        plastic: z.array(z.object({
-            name: z.string(),
-            description: z.string().optional().nullable(),
-            colours: z.array(z.object({
-                colour: z.string(),
-                id: z.string().optional().nullable()
-            })).optional().nullable(),
-            id: z.string().optional().nullable(),
-            blockName: z.string().optional().nullable(),
-            blockType: z.literal("plastic")
-        })),
-        layerHeight: z.number(),
-        infill: z.number()
-    }),
-    updatedAt: z.string(),
-    createdAt: z.string()
-});
-
-export const userSchema: z.ZodSchema<User> = z.lazy(() => z.object({
-    id: z.string(),
-    emailVerified: z.string().optional().nullable(),
-    name: z.string(),
-    image: z.string().optional().nullable(),
-    email: z.string(),
-    role: z.union([z.literal("customer"), z.literal("employee"), z.literal("admin"), z.literal("developer")]).optional().nullable(),
-    address: z.array(z.object({
-        line1: z.string(),
-        line2: z.string().optional().nullable(),
-        city: z.string(),
-        postalCode: z.string(),
-        country: z.string(),
-        id: z.string().optional().nullable()
-    })).optional().nullable(),
-    orders: z.array(z.union([z.string(), orderSchema])).optional().nullable(),
-    subscription: z.object({
-        plan: z.union([z.literal("none"), z.literal("plus")]).optional().nullable(),
-        status: z.union([z.literal("Active"), z.literal("PastDue"), z.literal("Unpaid"), z.literal("Paused")]).optional().nullable(),
-        stripeSubscriptionId: z.string().optional().nullable(),
-        currentPeriodStart: z.string().optional().nullable(),
-        currentPeriodEnd: z.string().optional().nullable()
-    }).optional(),
-    stripeCustomerId: z.string().optional().nullable(),
-    accounts: z.array(z.object({
-        provider: z.string(),
-        providerAccountId: z.string(),
-        type: z.union([z.literal("oidc"), z.literal("oauth"), z.literal("email"), z.literal("webauthn")]),
-        id: z.string().optional().nullable()
-    })).optional().nullable(),
-    updatedAt: z.string(),
-    createdAt: z.string(),
-    enableAPIKey: z.boolean().optional().nullable(),
-    apiKey: z.string().optional().nullable(),
-    apiKeyIndex: z.string().optional().nullable(),
-    resetPasswordToken: z.string().optional().nullable(),
-    resetPasswordExpiration: z.string().optional().nullable(),
-    salt: z.string().optional().nullable(),
-    hash: z.string().optional().nullable(),
-    _verified: z.boolean().optional().nullable(),
-    _verificationToken: z.string().optional().nullable(),
-    loginAttempts: z.number().optional().nullable(),
-    lockUntil: z.string().optional().nullable(),
-    password: z.string().optional().nullable(),
-    collection: z.literal("users")
-}));
-
-export const orderSchema: z.ZodSchema<Order> = z.lazy(() => z.object({
-    id: z.string(),
-    name: z.string(),
-    customer: z.union([z.string(), userSchema]),
-    prints: z.array(z.union([z.object({
-            product: z.union([z.string(), productSchema]),
-            quantity: z.number(),
-            price: z.number(),
-            completed: z.boolean().optional().nullable(),
-            id: z.string().optional().nullable(),
-            blockName: z.string().optional().nullable(),
-            blockType: z.literal("shopProduct")
-        }), z.object({
-            quote: z.union([z.string(), quoteSchema]),
-            model: z.object({
-                filename: z.string(),
-                filetype: z.union([z.literal("stl"), z.literal("3mf")]),
-                modelUrl: z.string(),
-                gcodeUrl: z.string()
-            }),
-            printingOptions: z.object({
-                preset: z.union([z.string().nullable(), presetSchema]).optional(),
-                layerHeight: z.number().optional().nullable(),
-                infill: z.number().optional().nullable(),
-                plastic: z.string(),
-                colour: z.string()
-            }),
-            time: z.string().optional().nullable(),
-            filament: z.number().optional().nullable(),
-            quantity: z.number(),
-            price: z.number(),
-            completed: z.boolean().optional().nullable(),
-            id: z.string().optional().nullable(),
-            blockName: z.string().optional().nullable(),
-            blockType: z.literal("customPrint")
-        })])),
-    payment: z.object({
-        provider: z.string().optional().nullable(),
-        stripeCustomerId: z.string().optional().nullable(),
-        stripeCheckoutSessionId: z.string().optional().nullable(),
-        stripePaymentIntentId: z.string().optional().nullable(),
-        stripeChargeId: z.string().optional().nullable(),
-        currency: z.string().optional().nullable(),
-        amount: z.number().optional().nullable(),
-        status: z.union([z.literal("awaiting-payment"), z.literal("paid"), z.literal("failed"), z.literal("refunded"), z.literal("partially-refunded"), z.literal("cancelled"), z.literal("expired"), z.literal("checkout-failed")]).optional().nullable(),
-        paidAt: z.string().optional().nullable(),
-        refunded: z.boolean().optional().nullable(),
-        refundedAmount: z.number().optional().nullable(),
-        refundedAt: z.string().optional().nullable(),
-        receiptUrl: z.string().optional().nullable()
-    }).optional(),
-    shipping: z.object({
-        shipmentId: z.string().optional().nullable(),
-        transactionId: z.string().optional().nullable(),
-        carrier: z.string().optional().nullable(),
-        service: z.string().optional().nullable(),
-        trackingNumber: z.string().optional().nullable(),
-        trackingUrl: z.string().optional().nullable(),
-        labelUrl: z.string().optional().nullable(),
-        labelPurchasedAt: z.string().optional().nullable(),
-        shippedAt: z.string().optional().nullable(),
-        deliveredAt: z.string().optional().nullable()
-    }).optional(),
-    dimensions: z.object({
-        length: z.number().optional().nullable(),
-        width: z.number().optional().nullable(),
-        height: z.number().optional().nullable(),
-        weight: z.number().optional().nullable()
-    }).optional(),
-    shippingAddress: z.object({
-        fullName: z.string().optional().nullable(),
-        line1: z.string().optional().nullable(),
-        line2: z.string().optional().nullable(),
-        city: z.string().optional().nullable(),
-        county: z.string().optional().nullable(),
-        postcode: z.string().optional().nullable(),
-        country: z.string().optional().nullable()
-    }).optional(),
-    pricing: z.object({
-        subtotal: z.number().optional().nullable(),
-        shipping: z.number().optional().nullable(),
-        tax: z.number().optional().nullable(),
-        total: z.number().optional().nullable()
-    }),
-    queue: z.number(),
-    status: z.object({
-        statuses: z.array(z.object({
-            stage: z.union([z.literal("in-queue"), z.literal("printing"), z.literal("packaging"), z.literal("shipped"), z.literal("cancelled")]),
-            timestamp: z.string(),
-            id: z.string().optional().nullable()
-        })).optional().nullable(),
-        currentStatus: z.union([z.literal("in-queue"), z.literal("printing"), z.literal("packaging"), z.literal("shipped"), z.literal("cancelled")])
-    }),
-    comments: z.string().optional().nullable(),
-    updatedAt: z.string(),
-    createdAt: z.string()
-}));
-
-export const quoteSchema: z.ZodSchema<Quote> = z.lazy(() => z.object({
-    id: z.string(),
-    model: z.object({
-        filename: z.string(),
-        filetype: z.union([z.literal("stl"), z.literal("3mf")]),
-        modelUrl: z.string().optional().nullable(),
-        gcodeUrl: z.string().optional().nullable()
-    }),
-    printingOptions: z.object({
-        preset: z.union([z.string().nullable(), presetSchema]).optional(),
-        layerHeight: z.number().optional().nullable(),
-        infill: z.number().optional().nullable(),
-        plastic: z.string(),
-        colour: z.string()
-    }),
-    customer: z.union([z.string(), userSchema]),
-    filament: z.number().optional().nullable(),
-    time: z.string().optional().nullable(),
-    price: z.number().optional().nullable(),
-    updatedAt: z.string(),
-    createdAt: z.string()
-}));
